app/services/dependencies.py
from abc import ABC, abstractmethod
from io import BytesIO
from typing import List
import os
from copy import deepcopy

from PIL import Image
from PyPDF2 import PdfMerger
import pytesseract

from app.utils import Graph
from llama_index.core import PropertyGraphIndex
from llama_parse import LlamaParse
from llama_index.core.schema import Document, TextNode

__all__ = [
    "ABC", "abstractmethod", "BytesIO", "List", "os", "deepcopy",
    "Graph", "Image", "PdfMerger", "pytesseract", 
    "PropertyGraphIndex", "LlamaParse", "Document", "TextNode"
]