requirements.txt
spacy==3.7.6
fastapi==0.115.4
uvicorn==0.32.0
transformers==4.46.0
accelerate==1.0.1
torch==2.4.0
llama-index==0.11.19
llama-index-core==0.11.19
llama-index-postprocessor-flag-embedding-reranker==0.2.0
llama-index-graph-stores-neo4j==0.3.5
llama-index-llms-huggingface==0.3.5
llama-index-embeddings-huggingface==0.3.1
llama-parse==0.5.12
fastcoref==2.1.6
ipywidgets==8.1.5
nltk>=3.8
FlagEmbedding==1.2.11
python-dotenv==1.0.1
ocrmypdf==15.4.4
PyPDF2==3.0.1
pytesseract==0.3.13